+++ src/three/IKRootsHelper.d.ts
@@ -6,7 +6,6 @@ export class IKRootsHelper extends Group {
 	constructor( roots : Array<Frame> );
 
 	setJointScale( scale : Number ) : void;
-	setResolution( width : Number, height : Number ) : void;
 	updateStructure() : void;
 	dispose() : void;