+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.0.5] - 2025-07-18
 ### Fixed
 - Adjust package version requirements.
 - Set the type and module entry in the package.
+- Updated the project to work with three.js r159 and up.
+
+### Changed
+- Removed `IKRootsHelper.setResolution`.
 
 ## [0.0.4] - 2022-11-21
 ### Added